core/requirements.txt
datasets==2.20.0
faiss_cpu==1.8.0.post1
sentence_transformers==3.0.1
